requirements.txt
cryptography==41.0.2
pycryptodome==3.18.0